+++ src/create/index.js
@@ -9,6 +9,7 @@ const path = require("path");
 const inquirer = require("inquirer");
 const render = require("consolidate").handlebars.render;
 const glob = require("glob").sync;
+var ui = new inquirer.ui.BottomBar();
 
 const { fail } = require("../utils");
 
@@ -18,14 +19,19 @@ const templates = glob(path.join(__dirname, "*.template")).map(f => path.parse(f
  * Yargs command
  */
 module.exports = {
-	command: "create <module>",
-	describe: `Create a Moleculer module (${templates.join(",")})`,
+	command: ["create service", "create service <name>"],
+	describe: `Create a Moleculer service (${templates.join(",")})`,
+	builder(yargs) {
+		yargs.options({
+			"typescript": {
+				describe: "Create service for typescript",
+				type: "boolean",
+				default: false
+			}
+		});
+	},
 	handler(opts) {
-		if (opts.module.toLowerCase() == "service")
-			return addService(opts);
-		else {
-			fail("Invalid module type. Available modules: " + templates.join(", "));
-		}
+		return addService(opts);
 	}
 };
 
@@ -38,39 +44,54 @@ module.exports = {
  */
 function addService(opts) {
 	let values = Object.assign({}, opts);
+	const _typescript = values.typescript ? true : false;
+	const [action, name] = values._;
 
 	return Promise.resolve()
 		.then(() => {
-			return inquirer.prompt([
-				{
-					type: "input",
-					name: "serviceFolder",
-					message: "Service directory",
-					default: "./services",
-					validate(input) {
-						if (!fs.existsSync(path.resolve(input)))
-							return `The '${input}' directory is not exists! Full path: ${path.resolve(input)}`;
-
-						return true;
+			const answers_options = [		{
+				type: "input",
+				name: "serviceFolder",
+				message: "Service directory",
+				default: "./services",
+				async validate(input) {
+					if (!fs.existsSync(path.resolve(input))){
+						ui.log.write(`The  ${input} doesn't exists!`)
+						fail("Aborted");
 					}
-				},
-				{
+					return true
+				}
+
+			}];
+
+
+			if(!name)
+				answers_options.push({
 					type: "input",
 					name: "serviceName",
 					message: "Service name",
 					default: "test"
-				}
-			]).then(answers => {
-				Object.assign(values, answers);
+				});
 
-				const newServicePath = path.join(values.serviceFolder, values.serviceName + ".service.js");
-				values.newServicePath = newServicePath;
+			return inquirer.prompt(answers_options).then(answers => {
+				answers.name = answers.serviceName;
+				answers.serviceName = answers.serviceName || name ;
+				answers.serviceName = answers.serviceName.replace(/[^\w\s]/gi, '-');
+
+				answers.className = answers.serviceName.replace(/(\w)(\w*)/g,
+					function(g0,g1,g2){ return g1.toUpperCase() + g2.toLowerCase();})
+					.replace(/[^\w\s]/gi, '');
+
+				Object.assign(values, answers);
+				const { serviceFolder , serviceName  } = values;
+				const file_name = `${serviceName.toLowerCase()}.service${_typescript ?".ts" :".js"}`;
+				const newServicePath =  path.join(serviceFolder, `${serviceName.toLowerCase()}.service${_typescript ?".ts" :".js"}`);
 
 				if (fs.existsSync(newServicePath)) {
 					return inquirer.prompt([{
 						type: "confirm",
 						name: "sure",
-						message: "The file is exists! Overwrite?",
+						message: `The file ${file_name} already exists! Do you want to overwrite it?`,
 						default: false
 					}]).then(({ sure }) => {
 						if (!sure)
@@ -80,17 +101,19 @@ function addService(opts) {
 			});
 		})
 		.then(() => {
-			const templatePath = path.join(__dirname, "service.template");
-			const template = fs.readFileSync(templatePath, "utf8");
 
-			return new Promise((resolve, reject) => {
-				render(template, values, function (err, res) {
+			const templatePath = _typescript ? path.join(__dirname, "typescript.service"):path.join(__dirname, "service.template");
+			const template = fs.readFileSync(templatePath, "utf8");
+			return new Promise( (resolve, reject) => {
+				render(template, values, async function (err, res) {
 					if (err)
 						return reject(err);
 
-					const { newServicePath } = values;
+					const { serviceFolder , serviceName  } = values;
+					const newServicePath =  path.join(serviceFolder, `${serviceName.toLowerCase()}.service${_typescript ?".ts" :".js"}`);
+
 					console.log(`Create new service file to '${newServicePath}'...`);
-					fs.writeFileSync(path.resolve(newServicePath), res, "utf8");
+					fs.writeFileSync(path.resolve(`${newServicePath}`), res, "utf8");
 
 					resolve();
 				});
@@ -99,4 +122,8 @@ function addService(opts) {
 
 		// Error handler
 		.catch(err => fail(err));
+}
+
+function capitalizeFirstLetter(string) {
+	return string[0].toUpperCase() + string.slice(1);
 }