+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="0.8.0"></a>
+# 0.8.0 (2021-01-08)
+
+## Breaking changes
+Updated `moleculer-repl` to 0.7.0. [Changelog](https://github.com/moleculerjs/moleculer-repl/releases/tag/v0.7.0)
+
+## Changes
+- update deps & audit fix
+
 --------------------------------------------------
 <a name="0.7.1"></a>
 # 0.7.1 (2020-04-09)